requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
langdetect==1.0.9
multidict==6.0.4
openai==0.27.8
pydub==0.25.1
python-dotenv==1.0.0
python-telegram-bot==20.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.0
tqdm==4.65.0
urllib3==2.0.3
yarl==1.9.2
